+++ crates/modor_graphics_new2/src/resources/material.rs
@@ -1,5 +1,4 @@
 use crate::gpu_data::uniform::Uniform;
-use crate::resources::mesh::MeshKey;
 use crate::resources::shader::ShaderKey;
 use crate::resources::texture::TextureRegistry;
 use crate::{
@@ -7,35 +6,35 @@ use crate::{
     Texture,
 };
 use modor::{Query, Single, SingleMut};
+use modor_math::Vec2;
 
 pub(crate) type MaterialRegistry = ResourceRegistry<Material>;
 
-// TODO: add texture position and size params
-
 // performance impact if lots of materials
 #[must_use]
-#[derive(Debug)]
+#[derive(Component, Debug)]
 pub struct Material {
     pub color: Color,
     pub texture_key: Option<ResourceKey>,
+    pub texture_position: Vec2,
+    pub texture_size: Vec2,
+    pub(crate) shader_key: ResourceKey,
     key: ResourceKey,
-    shader_key: ResourceKey,
-    mesh_key: ResourceKey,
     uniform: Option<Uniform<MaterialData>>,
     is_transparent: bool,
     old_is_transparent: bool,
 }
 
-#[component]
+#[systems]
 impl Material {
     const BINDING: u32 = 0;
 
-    pub fn rectangle(key: impl IntoResourceKey) -> Self {
-        Self::new(key, ShaderKey::Rectangle)
+    pub fn new(key: impl IntoResourceKey) -> Self {
+        Self::new_internal(key, ShaderKey::Default)
     }
 
     pub fn ellipse(key: impl IntoResourceKey) -> Self {
-        Self::new(key, ShaderKey::Ellipse)
+        Self::new_internal(key, ShaderKey::Ellipse)
     }
 
     pub fn with_color(mut self, color: Color) -> Self {
@@ -48,13 +47,24 @@ impl Material {
         self
     }
 
-    fn new(key: impl IntoResourceKey, shader_key: ShaderKey) -> Self {
+    pub fn with_texture_position(mut self, position: Vec2) -> Self {
+        self.texture_position = position;
+        self
+    }
+
+    pub fn with_texture_size(mut self, size: Vec2) -> Self {
+        self.texture_size = size;
+        self
+    }
+
+    fn new_internal(key: impl IntoResourceKey, shader_key: ShaderKey) -> Self {
         Self {
             color: Color::WHITE,
             texture_key: None,
+            texture_position: Vec2::ZERO,
+            texture_size: Vec2::ONE,
             key: key.into_key(),
             shader_key: shader_key.into_key(),
-            mesh_key: MeshKey::Rectangle.into_key(),
             uniform: None,
             is_transparent: false,
             old_is_transparent: false,
@@ -66,10 +76,8 @@ impl Material {
         if let Some(module) = module {
             let data = MaterialData {
                 color: self.color.into(),
-                texture_part_position: [0., 0.],
-                texture_part_size: [1., 1.],
-                has_texture: 0,
-                _padding: [0., 0., 0.],
+                texture_part_position: [self.texture_position.x, self.texture_position.y],
+                texture_part_size: [self.texture_size.x, self.texture_size.y],
             };
             if let Some(uniform) = &mut self.uniform {
                 if data != **uniform {
@@ -119,14 +127,6 @@ impl Material {
             .as_ref()
             .expect("internal error: material uniform not initialized")
     }
-
-    pub(crate) fn shader_key(&self) -> &ResourceKey {
-        &self.shader_key
-    }
-
-    pub(crate) fn mesh_key(&self) -> &ResourceKey {
-        &self.mesh_key
-    }
 }
 
 impl Resource for Material {
@@ -149,6 +149,4 @@ pub(crate) struct MaterialData {
     pub(crate) color: [f32; 4],
     pub(crate) texture_part_position: [f32; 2],
     pub(crate) texture_part_size: [f32; 2],
-    pub(crate) has_texture: u32,
-    pub(crate) _padding: [f32; 3],
 }