+++ crates/modor_graphics_new2/src/resources/mod.rs
@@ -1,7 +1,6 @@
 pub(crate) mod camera;
 pub(crate) mod material;
 pub(crate) mod mesh;
-pub(crate) mod render_target;
 pub(crate) mod shader;
 pub(crate) mod texture;
 pub(crate) mod window;