+++ CHANGELOG.md
@@ -4,10 +4,10 @@
 
 ### Added
 
-- Entity type definition
+- Entity, component and system definition
 - Sequential and parallel system execution
 - Physics module to handle transforms, relative transforms, dynamic properties and colliders
-- Graphics module to draw shapes (rectangles, ellipses), with customization (color, texture)
+- Graphics module to draw models (rectangles, ellipses), with material (color, texture)
 - Input module to access keyboard, mouse, touch and gamepad states
 - Asynchronous job utilities
 - Support for Windows, Linux, macOS, Android and WebAssembly